notification-system/requirements.txt
pika==1.1.0
python-dotenv